Connect sqlite3 DB
