+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "asdf",
+  "name": "recommended-vod",
   "version": "0.0.1",
   "description": "",
   "author": "",
