+++ src/contents/contents.module.ts
@@ -1,9 +1,12 @@
 import { Module } from '@nestjs/common';
-import { ContentsService } from './contents.service';
+import { ContentService } from './contents.service';
 import { ContentsController } from './contents.controller';
-
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Content, Category, Genre } from './entities/content.entity';
+import { VOD, Platform } from '../platforms/entities/platform.entity';
 @Module({
+  imports: [TypeOrmModule.forFeature([Content, Category, Genre])],
   controllers: [ContentsController],
-  providers: [ContentsService],
+  providers: [ContentService],
 })
 export class ContentsModule {}
