+++ src/database.providers.ts
@@ -1,0 +1,21 @@
+// import { DataSource } from 'typeorm';
+
+// export const databaseProviders = [
+//   {
+//     provide: 'DATA_SOURCE',
+//     useFactory: async () => {
+//       const dataSource = new DataSource({
+//         type: 'sqlite',
+//         // host: 'localhost',
+//         // port: 3306,
+//         // username: 'root',
+//         // password: 'root',
+//         database: 'r-vod.sqlite3',
+//         entities: [__dirname + '/../**/*.entity{.ts,.js}'],
+//         synchronize: true,
+//       });
+
+//       return dataSource.initialize();
+//     },
+//   },
+// ];
